+++ test/string_test.exs
@@ -27,6 +27,29 @@ defmodule Moar.StringTest do
     end
   end
 
+  describe "pluralize" do
+    test "returns the string when the count is 1 or -1" do
+      assert Moar.String.pluralize(1, "fish", "fishies") == "fish"
+      assert Moar.String.pluralize(-1, "fish", "fishies") == "fish"
+    end
+
+    test "returns the pluralized string when the count is not 1 or -1" do
+      assert Moar.String.pluralize(2, "fish", "fishies") == "fishies"
+      assert Moar.String.pluralize(0, "fish", "fishies") == "fishies"
+      assert Moar.String.pluralize(-2, "fish", "fishies") == "fishies"
+    end
+
+    test "accepts a pluralizer function" do
+      pluralizer = fn s -> s <> "ies" end
+
+      assert Moar.String.pluralize(2, "fish", pluralizer) == "fishies"
+      assert Moar.String.pluralize(1, "fish", pluralizer) == "fish"
+      assert Moar.String.pluralize(0, "fish", pluralizer) == "fishies"
+      assert Moar.String.pluralize(-1, "fish", pluralizer) == "fish"
+      assert Moar.String.pluralize(-2, "fish", pluralizer) == "fishies"
+    end
+  end
+
   describe "secure_compare/2" do
     test "compares binaries securely" do
       assert Moar.String.secure_compare(<<>>, <<>>)